Libtorrent upgrade and cross-build, WIP
Libtorrent upgraded from RC to release. Switched to the `configure` version of the build as it seems more flexible, transparent and self-contained.

Instead of using the system version of the Boost library we're now fetching it from the web and building a static version of it. This is slower, but it allows us to use a known-good version of Boost and it solves runtime failures arising when the system Boost version doesn't match the one we've been building against. And Boost no longer needs to be installed for mm2 to run.

And we've added code to cross-compile Boost and libtorrent for ARM, using the Docker-based `cross` crate.

The MacOS build will probably fail, going to test and patch it in a separate commit.

#265, #296