Start linking the mm2 binary for Android
Fixes the problems encountered during ARM7 cross-compilation with `cross`, allowing us to test linking the MM1 libraries.

The source code is mounted read-only in the `cross` build, so the generated bindings had to be moved into a different dir. Which is a good thing, really: having the generated headers in the source was causing trouble whenever the sources were copied between different OS platforms.

#296