Do not expose bindgen to threading
In the previous commit (18a8607) I have removed some of pthread types from the headers because these proved problematic for the cross-compiling bindgen. (Using pthread types in public headers is best to be avoided, IMHO, them not being the most portable). And I've missed some of the compilation errors that surfaced in a clean build. This commit contains the fallout, moving the pthread-using types from the headers and into C files.

#296