fix nonce locking issue
The nonce lock was only acquired while polling for the latest nonce. We
weren't waiting to send the transaction that uses this nonce nor till
the RPCs get these nonces updated.

This commit significantly increases the time taken in `test_nonce_lock` test (now around 2 minutes, but it should be way less with a dockerized geth node as RPC backend?).

Might relate to #610 since the documented way of acquiring and releasing
the nonce lock in that issue wasn't the same as the one in used before this commit