revert to trying connections in parallel in background task (to some extent)
so looks like trying connections sequntially introduced a long wait time till we got a valid working connection. this happens when the list of electrums are bunch of faulty ones and the last of them is working fine.
this didn't hurt the correctness in anyway, but made it kind of slow to bootstrap (finding the first working connection), and since tests doesn't try requests forever (forever-waiting until the request is successful is only used on cfg(not(test)) environments), this meant a lot of tests will fail to find a working connection soon enough and will panic on the first request.
we might want to find a strong fix for this.
In this commit, however, the fix to mitigate this issue is to re-intoduce concurrency to the point that the number of open connections at any time is less than or equal `max_connected` specified in the enable request. For multiple policy (min=1, max=len(servers)), this is just trying all the servers concurrently so that's fast, the multiple policy is the one used by default and in tests