fix the qrc20 issue on CI
So the issue was, we didn't have the = when comparing the number of maintained connections with min_connected.
But since we never maintain more then min_connected, this means the check without the = will always be false thus skiping the sleep code in the background task.
The background task was basically looping non-stop and degrading the performance of all other async tasks running on the runtime.
Very slow tests led to slow connections and timeouting everytime.