refactor expirable map
This refactors expirable map to make it so that deleting expired entries
is cheap (not looping over the whole map, but looping only over the
known expired entries).

This increases the memory requirements though because the key of type
`K` is now cloned an extra time to be stored in the inverse map.

A solution for the potentially expensive cloning is to wrap the key in
an Arc, but, looking at the usage of expirable map, it looks like most
(actually all) of the times, the keys are integer IDs, which are cheaply
cloned/copied. So leaving this as a direct clone for now until the need
arises.