let the responses map in eletrum connections be expriable
otherwise responses which timeout are never deleted :O

Also we now send electrum_request_multi in parallel, the first response
is considered and the others are discarded. This can make up end in a
case where one connection had the response notifier registered in the
responses map but the request was never sent, thus a response will never
be received, thus this entry will never be removed.