don't use block_on on mocked methods
because we now use tokio's block_on to wait for this method and not futures' wait. this means we will block_on the same runtime recurrsively and panic. So instead of block_on in mock methods, fut.boxed() like the original methods.

the difference between the two impls of the mock is that
the first one ran synchronously and returned an immediately ready future,
the second one creates an non-ready future to be ran asynchronously by the caller.