bug: don't disconnect the connection we want to connect while version checking
when doing version checking, the connection isn't marked as maintained yet, this means flagging the not needed will disconnect it immediately.
a fix for this is to keep a state of `Connecting` for example to avoid disconnecting a connection which is in progress, eaiser solution is to not ask for disconnection for connections that we didn't force connect.
so only if we caused the connection to connect will we call for it to be not needed anymore