disable the coin if the balance query fails
otherwise the GUI will get an error and think the coin wasn't enabled successfuly but it was. so disabling the coin to keep consistent with the GUIs expectionations.
one solution would be to omit the balance field if balance querying fialed, but we would need the GUI to account for that then