run rpc futures in a tokio thread till completion
when ran directly inside hyper's service function, it might get aborted mid-way when a client disconnects, leaving the future in complete.

this is an issue as we might have some code that need to be executed atomically inside any part of the code base that could be triggered by an RPC request. if the client disconnects, such a function will abort on the next await call, leaving us with non-atomic state (e.g. partial update for a map and its inverse).