cap the electrum request to 5s
the balance query by default could take up to 40s:
20s to establish a connnection with an electrum server &
20s to perform the electrum request.

the old electrum rpc would timeout in 5s if the electrum servers aren't responsive, and it seems like the desktop GUI accounted only for such timeout. let's use 5s timeout internally for the electrum request to make sure we are compatible with GUIs.
note that this only caps the timeout for the enable request to 5s, but in normal operations the other configured timeouts are used