convert async mutexs to sync ones
this carried over from the usage of these fields in past impl. there is no reason to have these mutexes async. sync mutexs are better for performance.
one mutex is left async though (establishing_connection) to not cpu-block other threads when they are waiting on the same connection to be established.