suggestion for address and shared DBs
so this is kind of an on-the-fly migration to the new directories. with the migration here being copying the old data to the new directories.
for addresses, the data is copied redundantly for each and every address format and each address index that the coin is instantiated with. this way we will avoid losing old data that is relevant to that address (it's hard to filter what data is relevent to copy it over, so we copy EVERYTHING).
we keep the old DB as is as later enables might still need it. this is very inefficient and is open to suggetions.

for the shared DB, the path changed from .kdf/DB/SEEDxMAGIC_DERIVED_HASH to .kdf/DB/hd_wallets/SEEDxKDF_INTERNAL_PATH_DERIVED_HASH, because i really couldn't find any justification for using the magic derived hash. this migration could be done with a rename instead of a recursive copy. leaving it like so for now though