make my_swaps_dir local to address_dir
and a whole wave of ripple effect of things that use it. if we don't know the address_dir (e.g. in rpc methods they only provide uuid, and it's not even practicial to ask them to provide the address that hosted the swap), we use load_my_swaps_from_unknown_db, which will lookup the maker address (same as the address_dir) from the global db lookup table (this is not implemented yet, but a placeholder is in place)