refactor tx building code
i can't say whether this code is more clear/readable since i am thinking in my own bubble. somebody else need to look at this and see if it really made things clearer.

i tried to approach this in a semi-functional fashion. i didn't like the mutation of tx builder fields and the tracked sums in the middle of the computation.
i got rid of the sums, removed some other non-necessary fields and made it so all the methods called inside build() don't mutate the builder by any means, they only do some computation on the current state of the builder.
the builder is the sole mutator in the tx building process. but not much either. the builder just keep adding more inputs until a valid tx (tx_solution = Ok) is found. then the only mutation done after finding the correct
fee & change is to deduct the fee from some output if instructed and to add the change to the outputs list.

the core of the code logic is now really scoped in the solver function `get_change_and_fee`, which i find is better than splitting the core logic over many functions. a single function is easier to verify esp with fee interactions.

there is some redundency of computations in the code. but i find that more easier to grasp after all, as maintaining sum fields (and like) and keeping them in sync is a complicated and error prone task IMO.