specify the dbdir for migration functions
even though they are not used yet. And I am not really sure if we will need this migrations if we are good for this being a breaking change. any old un-migrated dbs should be abandoned anyways and we can start a new migration scheme from scratch now.

i am only adding this just because