adjusting the tx history storage (or actually breaking it)
the compilation is now (finally) broke :)

the work here is attempting to move from storing all the tx history in a single DB to storing the tx history in the DB of the address which each tx deals with, namely: the from & to addresses (yes the tx is stored in each single one of them).
the methods were changed so that they always provide the addresses which we should use for DB access (e.g. get_tx_from_history now needs a for_addresses which only these set of addresses DB will be looked for the internal_id the caller is looking for).

the method get_history wasn't adjusted and it's still a pain point since it needs filters and paging which doesn't work good with the split DB model we have currently adopted. The paging issue could be fixed by using a global DB to store the order of txs inserted and then we can have a proper paging. for the filtering part though, there doesn't seem to be a good solution.

also, each address DB should be inited before accessing the DB. That isn't covered here. I wish we could have only a single table for tx history instead of BTC_tx_history, BCH_tx_history, etc... and instead specify the coin in a column in the table. this way we can init all the tables we need for any address DB from the normal migration code. In a sense, every call to `ctx.address_db` will always return a fully migrated address DB with all the tables needed.