new swap kickstart model (orders yet to come, and will be similar)
i really wanted to kickstart the swap on coin activations: that is, when the set of required maker and taker coins are activated with the correct addresses we fire a swap kickstart.

but this couldn't be achieved here, because coins are enabled inside lp_coins.rs which can't call swap methods that reside in lp_swaps.rs. so the approach taken here is to have a watcher in lp_swaps.rs watch for any new swaps that can be kickstarted, such swaps are registered as kicksartable from lp_coins.rs when the coin is enabled.
this is efficient than swapning a thread for each swap and not knowing will it ever be kickstarted or not, but still not the best. the best is to have a reactive way of informing that coin x is activated and swap uuid could be kickstarted, ideally using a channel (or maybe the multi-channel streaming interface we have).