move NftCtx out of EthCoin
well, it looks like having this ctx inside ethcoin won't work. the nftctx serves a similar thing to tx history (we have tx history in hd_wallet/shared db) and isn't related to the coin functionality or swaps (at least for now since the coin doesn't really serve anything and also no swaps for nfts, but if we later need this db for swaps we should move it into address based db and not shared) so for now it should reside a shared area and not address specific area. we still need a new column now for the from&to/owner address in the nft db so we can know which address it is dealing with. this wasn't accounted for since we only had iguana activation for nfts.