review(sami): cache global and wallet dbs in MmCtx
this avoids re-computations that will result in the same output. this also removes the fallible method annotation from global and wallet db getters, since we know for sure they are initialized and have a copy of them.

note that the name MM2-Shared.db for the wallet db was changed to a simple wallet.db. Also for the async wallet db (KOMODEFI.db), it changes to wallet.db as well for simplicity (merging them/their tables together in the same DB file).