don't create the address directory when calling address_dir()
we call address_dir() and then join the retruned pathbuf with other directory(s), i.e. nesting more directories (see e.g. eth_trace_path() or ln_data_dir()). this means we really need to also create not just the address directory but the directories we nest on top.

this commit removes the directory creation in address_dir() and makes this method infallible, this simplifies it's usage. BUT, now the caller needs to create the address directory (or make sure it's created) and any nested directories needed on top before using that directory or accessing any files in it