adjust verify_p2pk_input_pubkey() to account for both pubkey formats
found in scriptPub. i.e. we will assume the public key in scriptpub was compressed and try to verify the signature, if we fail, we will assume it's uncompressed and try to verify the signature again. if either of these two operations succeed, verification succeeds, if both fail, verification fails.