review(dimxy): just query for the prev_output and check it against the expected pubkey
this simplifies things a lot as no verification is now needed. the down side is that we now query the rpc for each and every p2pk input and not just overwintered ones. this is prolly OK tho since p2pk inputs aren't that popular anyways.

ofc we assume the rpc server isn't lying