write swaps in address_db and not the old db
the gist of this commit lies in mm2_main::database::my_swaps. `fn insert_new_swap` now takes a connection and writes that swap to. this connection is passed from the caller, and it's the address_db that's derived from the maker coin address.
for old arch this essentially brings no change. the same db from dbdir() is still the one used in this case.