use xpubs for db queries that edit hd_account table
xpubs and account_ids are pretty much the same thing. the only difference is that xpub is more specific as it also encodes not just the account id, but also the purpose/coin/ sections of the derivaiton.
by using xpubs for altering/editing accounts, we make sure we don't mis edit an account that has a different purpose/coin/ fields but with the same ticker (like the confustion that happens when ticker purpose/coin/ changes but the ticker name remains the same).
said simply, performing the queries with xpubs is more like performing the queries with purpose/coin/account_id all specified instead of just account_id.