manually convert PrivKeyBuildPoilcy to the ETH counterpart in legacy activation
we used to convert it using `priv_key_policy.into()` and then do a check to rule out the bad activation policies not accepted by legacy activation.
changing it now to just do the conversion inline and reject bad activation policies from the beginning. this way we don't accidently add a new activation policty but forget to rule it out in legacy activation