fix failing tests
looks like serde won't regonize an object that has no protocol_data when the enum content is inlined as struct representation in the enum.
to curcumvent this without forcing us to use protocol_data everywhere, we let the enum hold its content in a tuple representation and let that tuple hold a struct that has the needed content.