make more sense of the error types
nosuchcoin and coinnotsupported one was throwing a custom error and one was embedding the coin ticker. both of them can just include the coin ticker as the error variant name is arleady pretty descriptive.
but since the request is coin-based, we don't really need to return the coin ticker, the requester implicilty knows that nosuchcoin means the coin they asked for does not exist.