add a bad accounts fixing step in enable_hd_wallet
this step is supposed to look for bad accounts and fix them to (replace them by) good accounts by setting the correct purpose and coin_types. a bad account is fixed into a good one only if the derived xpub matches the one stored in the bad account. if they don't match, this means the current purpose and coin_type of the coin don't match the ones who were used to create the bad account.

this commit has compilation errors for multiple reasons. but still separating it away to be coherent. fixing the compilation errors in the next commits